split loguru.hpp into loguru.hpp and loguru.cpp
